wishes module

+++ src/wishes/wishes.controller.ts
@@ -13,6 +13,8 @@ import { WishesService } from './wishes.service';
 import { CreateWishDto } from './dto/create-wish.dto';
 import { UpdateWishDto } from './dto/update-wish.dto';
 import { JwtGuard } from '../guards/auth.guard';
+import { ServerException } from '../exceptions/server.exception';
+import { ErrorCode } from '../exceptions/error-codes';
 
 @Controller('wishes')
 export class WishesController {
@@ -24,23 +26,95 @@ export class WishesController {
     return this.wishesService.create(req.user, createWishDto);
   }
 
+  @UseGuards(JwtGuard)
+  @Get('/last')
+  async findLast() {
+    const last = await this.wishesService.findLast();
+
+    if (!last.length) {
+      throw new ServerException(ErrorCode.WishNotFound);
+    }
+    return last[0];
+  }
+
+  @UseGuards(JwtGuard)
+  @Get('/top')
+  async findTop() {
+    const last = await this.wishesService.findLast();
+
+    if (!last.length) {
+      throw new ServerException(ErrorCode.WishNotFound);
+    }
+    return last[0];
+  }
+
+  @UseGuards(JwtGuard)
+  @Get(':id')
+  async findOne(@Param('id') id: string) {
+    const wish = await this.wishesService.findOne(+id);
+    if (!wish) {
+      throw new ServerException(ErrorCode.WishNotFound);
+    }
+    return wish;
+  }
+
+  @UseGuards(JwtGuard)
+  @Patch(':id')
+  async update(
+    @Req() req,
+    @Param('id') id: string,
+    @Body() updateWishDto: UpdateWishDto,
+  ) {
+    const wish = await this.wishesService.findOne(+id);
+    if (!wish) {
+      throw new ServerException(ErrorCode.WishNotFound);
+    }
+
+    if (wish.owner.id !== req.user.id) {
+      throw new ServerException(ErrorCode.WishCanNotEdit);
+    }
+
+    return this.wishesService.update(+id, updateWishDto);
+  }
+
   @Get()
   findAll() {
     return this.wishesService.findAll();
   }
 
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.wishesService.findOne(+id);
-  }
-
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateWishDto: UpdateWishDto) {
-    return this.wishesService.update(+id, updateWishDto);
-  }
-
+  @UseGuards(JwtGuard)
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.wishesService.remove(+id);
+  async remove(@Req() req, @Param('id') id: string) {
+    const wish = await this.wishesService.findOne(+id);
+    if (!wish) {
+      throw new ServerException(ErrorCode.WishNotFound);
+    }
+
+    if (wish.owner.id !== req.user.id) {
+      throw new ServerException(ErrorCode.WishCanNotDelete);
+    }
+
+    await this.wishesService.remove(+id);
+
+    return wish;
+  }
+
+  @UseGuards(JwtGuard)
+  @Post(':id/copy')
+  async copy(@Req() req, @Param('id') id: string) {
+    const wish = await this.wishesService.findOne(+id);
+    if (!wish) {
+      throw new ServerException(ErrorCode.WishNotFound);
+    }
+
+    const { name, image, description, link, price } = wish;
+
+    return await this.wishesService.copy(req.user, {
+      name,
+      image,
+      description,
+      link,
+      price,
+    });
   }
 }
